pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "docling>=2.74.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "markdown2>=2.5.4",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "weasyprint>=68.1",
    "pypdf>=5.0.0",
    "pillow>=10.0",
]

[tool.setuptools.packages.find]
where = ["."]
